+++ src/betterdiscord/stores/settings.ts
@@ -76,7 +76,7 @@ export default new class SettingsManager extends Store {
         };
         if (options.manager) section.manager = options.manager;
         if (onClick) section.clickListener = onClick;
-        if (element) section.element = element instanceof DiscordModules.React.Component ? () => DiscordModules.React.createElement(element, {}) : typeof (element) == "function" ? element : () => element;
+        if (element) section.element = element instanceof (DiscordModules.React?.Component || Object) ? () => DiscordModules.React.createElement(element as any, {}) : typeof (element) == "function" ? element : () => element;
         this.panels.push(section);
     }