+++ src/betterdiscord/types/discord/webpack.d.ts
@@ -5,6 +5,7 @@ export interface Require {
     c: Record<PropertyKey, Module>;
     m: Record<PropertyKey, RawModule>;
     e(id: PropertyKey): Promise<unknown>;
+    l(url: string, onLoad: (event: Event) => void, key: string, id: string): void;
 }
 
 export interface Module<T extends any = any> {
@@ -35,9 +36,11 @@ export type Options = {
     fatal?: boolean;
     firstId?: PropertyKey;
     cacheId?: string | null;
+    declarationFilter?: ExportedOnlyFilter;
 };
 
 export type MangledOptions = Options & {
+    mapDeclarations?: boolean;
     useDeclarations?: boolean;
 };
 
@@ -45,6 +48,7 @@ export type BulkQueries = Options & {
     filter: Filter,
     all?: boolean,
     map?: Record<string, ExportedOnlyFilter>;
+    mapDeclarations?: boolean;
 };
 export type WithKeyOptions = Options & {
     target?: any;