+++ src/betterdiscord/webpack/searching.ts
@@ -1,5 +1,5 @@
 import type {Webpack} from "discord";
-import {getDefaultKey, makeException, shouldSkipModule, wrapFilter} from "./shared";
+import {getDeclaration, getDefaultKey, makeException, shouldSkipModule, wrapFilter} from "./shared";
 import {webpackRequire} from "./require";
 import WebpackCache from "./cache";
 
@@ -9,6 +9,7 @@ export function getMatched<T>(module: Webpack.Module<any>, filter: Webpack.Filte
     if (shouldSkipModule(module.exports)) return;
 
     if (filter(module.exports, module, module.id)) {
+        if (options.declarationFilter) return getDeclaration(module, options.declarationFilter);
         return raw ? module as T : module.exports;
     }
 
@@ -26,6 +27,7 @@ export function getMatched<T>(module: Webpack.Module<any>, filter: Webpack.Filte
         if (shouldSkipModule(exported)) continue;
 
         if (filter(exported, module, module.id)) {
+            if (options.declarationFilter) return getDeclaration(module, options.declarationFilter);
             if (!defaultExport && defaultKey === key) {
                 return module.exports;
             }
@@ -90,7 +92,11 @@ export function getAllModules<T extends unknown[]>(filter: Webpack.Filter, optio
         if (shouldSkipModule(module.exports)) continue;
 
         if (filter(module.exports, module, module.id)) {
-            modules.push(raw ? module : module.exports);
+            if (options.declarationFilter) {
+                const declared = getDeclaration(module, options.declarationFilter);
+                if (declared) modules.push(declared);
+            }
+            else {modules.push(raw ? module : module.exports);}
         }
 
         if (!searchExports && !searchDefault) continue;
@@ -107,6 +113,12 @@ export function getAllModules<T extends unknown[]>(filter: Webpack.Filter, optio
             if (shouldSkipModule(exported)) continue;
 
             if (filter(exported, module, module.id)) {
+                if (options.declarationFilter) {
+                    const declared = getDeclaration(module, options.declarationFilter);
+                    if (declared) modules.push(declared);
+                    continue;
+                }
+
                 if (!defaultExport && defaultKey === key) {
                     modules.push(module.exports);
                     continue;