+++ src/betterdiscord/webpack/utilities.ts
@@ -3,7 +3,7 @@
 import type {Webpack} from "discord";
 import {bySource} from "./filter";
 import {getModule} from "./searching";
-import {getDefaultKey, makeException, shouldSkipModule, wrapFilter} from "./shared";
+import {getDefaultKey, makeException, shouldSkipModule, wrapFilter, getDeclaration} from "./shared";
 import {webpackRequire} from "./require";
 import WebpackCache from "./cache";
 import {mapObject} from "@utils/object";
@@ -42,12 +42,12 @@ export function getMangled<T extends object>(
         filter = bySource(filter);
     }
 
-    options.raw ??= options.useDeclarations ?? false;
+    options.raw ??= options.mapDeclarations ?? options.useDeclarations ?? false;
 
     let module = typeof filter === "number" ? getById(filter, options) : getModule<any>(filter, options);
     if (!module) return {} as T;
 
-    if (options.raw) module = module[options.useDeclarations ? "declarations" : "exports"];
+    if (options.raw) module = module[options.mapDeclarations || options.useDeclarations ? "declarations" : "exports"];
 
     return mapObject(module, mappers);
 }
@@ -56,6 +56,8 @@ export function bulkGetMatched<T>(module: Webpack.Module<any>, options: Webpack.
     const {filter, defaultExport = true, searchExports = false, searchDefault = true, raw = false, map} = options;
 
     if (filter(module.exports, module, module.id)) {
+        if (options.declarationFilter) return getDeclaration(module, options.declarationFilter);
+        if (options.mapDeclarations && options.map) return mapObject(module.declarations, options.map) as T;
         const trueItem = map ? mapObject(module.exports, map) : raw ? module : module.exports;
         return trueItem;
     }
@@ -71,6 +73,9 @@ export function bulkGetMatched<T>(module: Webpack.Module<any>, options: Webpack.
         if (shouldSkipModule(exported)) continue;
 
         if (filter(exported, module, module.id)) {
+            if (options.declarationFilter) return getDeclaration(module, options.declarationFilter);
+            if (options.mapDeclarations && options.map) return mapObject(module.declarations, options.map) as T;
+
             let value: any;
 
             if (!defaultExport && defaultKey === key) {